Create a Wordcloud

create_wordcloud produces a wordcloud with the lyrics of songs in a playlist that are already stored within the geniusR package.

create_wordcloud(data = data, playlist, stop_vector = F,
  del_file = T)

Arguments

data
a dataframe of a user's playlists

playlist
a character string

stop_vector
a boolean

del_file
a boolean

Value

a wordcloud generated from rquery

Details

This function takes in a string of a playlist, a stop_vector that can remove certain common words like "yeah" and "like" (as well as any other words the user find as irrelevent), and del_file which is a boolean value that asserts whether the user wants to delete the output text file that is generated within the function and that stores all the lyrics that are accumulated for the purposes of generating the word cloud.

Examples

data(christmas_playlists)
create_wordcloud(data = christmas_playlists, "Christmas Classics", c("yeah", "like"), del_file = T)

#> Loading required package: geniusR

#> Loading required package: tm

#> Loading required package: NLP

#> 
#> Attaching package: ‘NLP’

#> The following object is masked from ‘package:ggplot2’:
#> 
#>     annotate

#> Loading required package: SnowballC

#> Loading required package: RColorBrewer

#> Loading required package: RCurl

#> Loading required package: bitops

#> Loading required package: XML

#> Error: object 'geniusr_song' not found