Definitions Dataframe

A dataframe containing the Spotify parameters in the Spotify API, their value type (int or float), and their description as it's shown in the Spotify API.

definitons

Format

A data frame with 13 rows and three variables The variables are as follows:

KEY

name of parameter

VALUE TYPE

type int or float

3
VALUE DESCRIPTION

description of parameter from Spotify API

Source

The data was sourced from here.

Examples

head(definitions)

#>              KEY VALUE TYPE
#> 1    duration_ms        int
#> 2            key        int
#> 3           mode        int
#> 4 time_signature        int
#> 5   acousticness      float
#> 6   danceability      float
#>                                                                                                                                                                                                                              VALUE DESCRIPTION
#> 1                                                                                                                                                                                                   The duration of the track in milliseconds.
#> 2                                                      The estimated overall key of the track. Integers map to pitches using standard Pitch Class notation . E.g. 0 = C, 1 = C♯/D♭, 2 = D, and so on. If no key was detected, the value is -1.
#> 3                                                                              Mode indicates the modality (major or minor) of a track, the type of scale from which its melodic content is derived. Major is represented by 1 and minor is 0.
#> 4                                                                                An estimated overall time signature of a track. The time signature (meter) is a notational convention to specify how many beats are in each bar (or measure).
#> 5                                                                                                                 A confidence measure from 0.0 to 1.0 of whether the track is acoustic. 1.0 represents high confidence the track is acoustic.
#> 6 Danceability describes how suitable a track is for dancing based on a combination of musical elements including tempo, rhythm stability, beat strength, and overall regularity. A value of 0.0 is least danceable and 1.0 is most danceable.