Get Top Songs From a User Playlist

playlist_top_songs produces a dataframe with the top 5 songs in a given playlist regarding the given parameter.

playlist_top_songs(data = data, playlist, param, top = 5, asc = T,
  kable = T, bg = "azure", text = "palevioletred")

Arguments

data
a dataframe of a user's playlists

playlist
a character string

param
a numeric variable in data

top
an integer

asc
a boolean

kable
a boolean

bg
a character string

text
a character string

Value

a dataframe

Details

This function takes in a string of a playlist and a string of parameter and returns a dataframe with the top 5 songs

Examples

data(christmas_playlists)
playlist_top_songs(data = christmas_playlists, "Christmas Jazz", valence, asc = T, kable = F, top = 10)

#> # A tibble: 10 x 2
#>    track_name                                                            valence
#>    <chr>                                                                   <dbl>
#>  1 Jingle Bells                                                            0.892
#>  2 We Wish You The Merriest                                                0.869
#>  3 Frosty The Snowman                                                      0.841
#>  4 Sleigh Ride                                                             0.829
#>  5 Here Comes Santa Claus (Right Down Santa Claus Lane) - Single Version   0.749
#>  6 Baby, It's Cold Outside (feat. Meghan Trainor)                          0.684
#>  7 Man With The Bag                                                        0.684
#>  8 What A Wonderful Time Of Year                                           0.635
#>  9 Little Jack Frost Get Lost                                              0.631
#> 10 Jingle Bell Rock (with The Patrick Williams Orchestra)                  0.626